+++ cextern/wcslib/C/wcsprintf.h
@@ -1,7 +1,7 @@
 /*============================================================================
 
-  WCSLIB 4.20 - an implementation of the FITS WCS standard.
-  Copyright (C) 1995-2013, Mark Calabretta
+  WCSLIB 4.21 - an implementation of the FITS WCS standard.
+  Copyright (C) 1995-2014, Mark Calabretta
 
   This file is part of WCSLIB.
 
@@ -22,10 +22,10 @@
 
   Author: Mark Calabretta, Australia Telescope National Facility, CSIRO.
   http://www.atnf.csiro.au/people/Mark.Calabretta
-  $Id: wcsprintf.h,v 4.20 2013/12/18 05:42:49 mcalabre Exp $
+  $Id: wcsprintf.h,v 4.21 2014/03/24 05:12:27 mcalabre Exp $
 *=============================================================================
 *
-* WCSLIB 4.20 - C routines that implement the FITS World Coordinate System
+* WCSLIB 4.21 - C routines that implement the FITS World Coordinate System
 * (WCS) standard.
 *
 * Summary of the wcsprintf routines